Update apache.txt
So this change makes the whitelist check a positive  instead of a negative.   Basic logic here: if the REMOTE_ADD or the X-Forwarded-For address match the whitelist, then continue with the content checking Rewrite rules to match the content of the c2 URIs.   If both checks fail then it 404s the request.